lab-8 提交

+++ kernel/fs/bitmap.c
@@ -27,12 +27,12 @@ void bitmap_free_block(uint32 block_num)
 
 }
 
-uint32 bitmap_alloc_inode()
+uint16 bitmap_alloc_inode()
 {
 
 }
 
-void bitmap_free_inode(uint32 inode_num)
+void bitmap_free_inode(uint16 inode_num)
 {
 
 }
